Avoid loading oh-my-zsh if already set
Because of `source ~/.zshrc`, the files in the folder `~/.oh-my-szh` were loaded
causing an error with the `autoload` command. We then make sure to source the
strict minimum (in our case asdf) and avoid to dowload again `oh-my-zsh` if it's
already set.
The updates of oh-my-zsh are done continuously and are safe.