Touch .civilcode.zshrc
When setting up the laptop for the first time, we have not run rcrc yet.
Sourcing the zshrc file fille then lead to an error because it could not find the civilcode.zshrc file.
Touching it will ensure that it always exist.

Running then a rcup later will ask to overwrite the file:

```
➜ rcup
overwrite /Users/nicolas/.aliases? [ynaq]
```